requirements.txt
absl-py
beautifulsoup4
bilix
py-cord[speed]
pytube
requests
StringProgressBar
urllib3